rename workbench to examples
