+++ src/main/java/org/matsim/pt2matsim/examples/MappingAnalysisExample.java
@@ -16,7 +16,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.pt2matsim.workbench;
+package org.matsim.pt2matsim.examples;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Network;
