+++ src/main/java/org/matsim/pt2matsim/examples/PTMapperShapesExample.java
@@ -1,4 +1,4 @@
-package org.matsim.pt2matsim.workbench;
+package org.matsim.pt2matsim.examples;
 
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.pt.transitSchedule.api.TransitSchedule;
